+++ html-formatter/javascript/src/main.tsx
@@ -4,7 +4,7 @@ import {
   EnvelopesQuery,
   FilteredResults,
 } from '@cucumber/react'
-import { Query as GherkinQuery } from '@cucumber/gherkin'
+import { Query as GherkinQuery } from '@cucumber/gherkin-utils'
 import { Query as CucumberQuery } from '@cucumber/query'
 import React from 'react'
 import ReactDOM from 'react-dom'